requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-extensions==2.0.7
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
psycopg2-binary==2.8.6
gunicorn==20.1.0
drf-yasg==1.17.1
python-dotenv==1.0.0
